+++ key-constraints/PrimaryKeyMaintainer.js
@@ -1,7 +1,7 @@
 "use strict";
 var CollectionMetaData = require("./CollectionMetaData");
 var DataCollection = require("../db-collections/DataCollection");
-/* PrimaryKeyMaintainer - helper for {@link DataCollection}
+/* PrimaryKeyMaintainer - helper for DataCollection
  * @author TeamworkGuy2
  */
 var PrimaryKeyMaintainer = (function () {
@@ -53,9 +53,9 @@ var PrimaryKeyMaintainer = (function () {
         this.reloadAll = reloadAll;
     }
     /** Manage (add/track) unique IDs for objects added to data collections
-     * @param {string} collectionName: the collection name (lower underscore case)
-     * @param {any[]} docs: the list of objects being added to the 'collectionName' collection
-     * @param {boolean} addGeneratedKeys: true to generate unique IDs for properties that use
+     * @param collectionName: the collection name (lower underscore case)
+     * @param docs: the list of objects being added to the 'collectionName' collection
+     * @param addGeneratedKeys: true to generate unique IDs for properties that use
      * auto-generated keys, false to just track newly added items to ensure unique IDs are maintained
      */
     PrimaryKeyMaintainer.prototype.manageKeys = function (collectionName, docs, addGeneratedKeys) {
@@ -174,7 +174,7 @@ var PrimaryKeyMaintainer = (function () {
         }
         return largestValues;
     };
-    /** creates an {@link DataCollection} if {@code dataColl} is null, otherwise returns {@code dataColl} unmodified */
+    /** creates an DataCollection if 'dataColl' is null, otherwise returns 'dataColl' unmodified */
     PrimaryKeyMaintainer.initMetaDataCollection = function (dataColl, collectionName, dataModel, modelFuncs, dbDataInst) {
         if (dataColl == null) {
             var collModel = dbDataInst.getModelDefinitions().addModel(collectionName, dataModel, modelFuncs);